+++ Source/Plugins/IO/IOFilters/YSChoiAbaqusReader.cpp
@@ -259,7 +259,7 @@ void YSChoiAbaqusReader::dataCheck()
         xpoints = tokens[1].toInt(&ok, 10);
         ypoints = tokens[2].toInt(&ok, 10);
         zpoints = tokens[3].toInt(&ok, 10);
-        size_t dims[3] = {xpoints, ypoints, zpoints};
+        size_t dims[3] = { static_cast<size_t>(xpoints), static_cast<size_t>(ypoints), static_cast<size_t>(zpoints) };
         m->setDimensions(dims);
         m->setOrigin(0, 0, 0);
       }
@@ -353,7 +353,7 @@ void YSChoiAbaqusReader::execute()
       ypoints = tokens[2].toInt(&ok, 10);
       zpoints = tokens[3].toInt(&ok, 10);
       totalpoints = xpoints * ypoints * zpoints;
-      size_t dims[3] = {xpoints, ypoints, zpoints};
+      size_t dims[3] = { static_cast<size_t>(xpoints), static_cast<size_t>(ypoints), static_cast<size_t>(zpoints) };
       m->setDimensions(dims);
       m->setOrigin(0, 0, 0);
